[zircon][libc] Update __sanitizer_memory_snapshot callback documentation
We hit deadlock when we take a snapshot where a callback waits on a
futex that would be woken by a suspended thread. This is possible if a
newly created thread is suspended in the middle of a call to printf and
the main thread attempts to call printf in a callback. Because the lock
for stdout is held by the suspended thread, the main thread will never
be able to print until it gets the lock back.

This type of situation is more an issue with how the user uses
__sanitizer_memory_snapshot. The callbacks should ideally just iterate
and read over memory regions provided to the callbacks and nothing more.
This updates the documentation so users are more aware of this.

Fixed: 98237
Change-Id: I6df6310d57c2659942fc73ddfcafa3dede955082
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671555
Reviewed-by: Roland McGrath <mcgrathr@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>