[zircon][sanitizer] Ensure memory snaposhots cover the pthread start_arg
We can end up in a situation where an lsan scan happens at the end of a
main thread, but the before the new thread is suspended and hasn't even
started. If so, then the argument provided to pthread_create won't be
scanned because the thread register for the new thread wasn't setup yet.
To fix this, we can iterate over the start_arg member in pthread structs
in the global thread list.

Bug: 66819
Change-Id: Ic940b8cc3676323f643daebe09664196f444aaee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666554
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>