[brcmfmac] Switch sdio over to network device
With core supporting network device buses this change switches the sdio
bus layer over to using the generic network device as a dataplane.
Additionally it switches over to using the new SDIO driver API for all
transfers. This is done by creating an internal VMO used for transfers
that do not pertain to the dataplane. This VMO is also needed for
transfers that occur before the network device has been initialized,
such as firmware download and initial configuration.

Bug: 79622
Test: Extensive long term performance tests, thousands of reboots to
verify correct connection on startup.

Change-Id: I80477e9d4b1ef2ae005bef598e72a44ce5266c37
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612803
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>