[security] Allow ambient executability for workstation elements.
This unblocks a runnerless developer workflow for Flutter on Fuchsia
by giving JIT Flutter apps a place to run.

We want a runnerless workflow because Dart has no ABI compatibility.
With a runner-based workflow this forces the Flutter runner and Flutter
components to be created using the same Dart SDK which is inconvenient
for OOT workflows.

The non-vendored component_manager_policy.json5 file is only used _eng builds so it should be fine for the allowlist entry to live in this file,
but we should investigate ways of defining our allowlist that explicitly
declare what configuration they're allowed to be used for.

Change-Id: I664a441bc723b1c24bd7ea882dd34857f6ee5335
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678962
Reviewed-by: Drew Fisher <zarvox@google.com>
Commit-Queue: Alexander Biggs <akbiggs@google.com>