[ffx][test] Move directory schema specifics into directory library
Moves all details about the structure of the directory output format
into the directory library. Before this change, the library contained
only the schemas for the JSON files contained in the output format, and
clients (run_test_suite and its integration tests) were expected to
understand the structure of the directory.
The transition to the stable output format changes the number of JSON
files present in the output format. Moving these details to a single
library simplifies this transition somewhat. It also gives us a
mechanism to specify which version of the format to use while we
migrate clients.

Bug: 81195
Testing: fx test //src/sys/run-test-suite

Change-Id: I37813fad004ef90543505f96fcd195f2ac85a690
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647349
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>