[dns] Prefer DHCP-discovered nameservers
In https://fxrev.dev/669854, we "promoted" DHCPv6 servers priority above
DHCP with the goal of keeping nameservers discovered over some address
family in the same priority. However, until https://fxbug.dev/48868 is
completed, networks that rely solely on DHCPv6 (and not also on SLAAC)
for address assignment won't be able to reach the v6 Internet.

Given that such scenarios are possible (and confusing), when a
DHCP-supplied configuration can work, it makes sense to prefer that.

Bug: 98201
Change-Id: Ie833a900391b16a238277f8fdc162f5a19ed4c2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678426
Commit-Queue: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>