[mdns] allow publication to outlive publishers
This CL fixes an issue in which A publication was being terminated
when the publisher that created it was closed. This required the
client to keep the ServiceInstancePublisher channel open to keep
publishers created with the PublishServiceInstance method from being
destroyed. Lifetime of publishers is more logically tied to the
lifetime of the ServiceInstancePublicationResponder channel used
by the publisher to talk to the client.

This CL gives publishers their own lifetime independent of their
creators. Lifetime is tied to the channel, so the client can stop
publication when desired.

Fixed: 99755

Test: new unit test
Change-Id: Id559e386769607dd1bb0cca9c352cf7c3aae7581
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678223
Reviewed-by: Nate Hurley <nahurley@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>