[brcmfmac] Fix tail length for SDIO frames
SDIO TX frames did not specify any tail length which caused a debug
assert to trigger when adding tail space for alignment. In practice this
is harmless because the TX buffers are guaranteed to be larger. But for
the sake of correctness and potential future changes it's better to
allocate this space.

Bug: 100248
Test: Manual, sim_device_test brcmfmac_sim_test brcmfmac_test
      wlan-drivers-components-tests brcmfmac_sdio_test
Change-Id: Ia27079f84ae90fb76b64aa9d3df05b488ffb665f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679122
Fuchsia-Auto-Submit: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>