[virtualization] Add FIDL API for device orchestrated connections
(Note: This is an alternative to fxr/678164).

This Connect2 API would be hosted by the device, and remove the need
for a host_vsock binary. Clients would call Listen to allow for guest
initiated connections, and Connect to initiate connections to a guest.

See go/out-of-process-vsock-device for discussion.

Change-Id: I65d7efef17e5ff722d40f583e3f36a7c70d0ac8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678662
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Hastings <dahastin@google.com>
API-Review: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>