[brcmfmac] Fix caching of SDIO block size
The block size for SDIO is cached for performance reasons since it's
accessed for every transfer that needs to be aligned to block size.
Unfortunately caching happened before the block size was actually set.
This lead to the cached value having the factory default block size. In
the past we used the factory default value so it was not a problem. As
part of the network device work the block size was reduced and the
mismatch between cached value and actual block size became apparent.

Fortunately this hasn't led to any issues since the actual block size
was lower, all that happened was that we aligned to a stricter value
than necessary.

Bug: 100319
Test: manual testing, fx test brcmfmac_test brcmfmac_sdio_test
Change-Id: Ibb0f694050eeff8693747158ea318241ab05ad60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679450
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>