[kernel][vm] Fix ZeroRange for pager-backed VMOs
A ZeroRange (ZX_VMO_OP_ZERO) on a pager-backed VMO emulates a VMO write
with zeros. Previously this would decommit pages in the VMO, regardless
of the VMO type, which is incorrect for pager-backed VMOs. An empty page
slot in a pager-backed VMO indicates true absence of content, not zero
content as is the case for anonymous VMOs.

If the ZeroRange applies to the range beyond supply_zero_offset, we can
efficiently zero pages without committing any new pages as the range
starting at supply_zero_offset is implicitly zero. This is an
optimization targeting zero'ing the tail end of the VMO.

Test: core-pager-writeback
Bug: 63989
Run-All-Tests: true
Change-Id: I14b64b7af45bded4080afded1f0bd0cafc28515c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677101
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>