[loader_service] Use FIDL wire format v2 in ldmsg
Currently, loader_svc_rpc sends a message header indicating it is using
FIDL wire format v1, however use of FIDL wire format v1 is deprecated
and most bindings have switched over to expecting v2 in most cases.

For the messages in the fuchsia.ldsvc.Loader protocol, the v1 and v2
wire format layouts are identical. The main difference is that the
header is not sending a flag that indicates that the message is a v2
wire format message.

This CL introduces a ldmsg_req_init_txn_header function to initialize
the transaction header and ensure header initialization logic is shared
between dynlink.c and loader_service_test.cc.

This CL was created in response to debugging a FIDL CL which broke
driver_host.

Bug: 38643
Change-Id: I3b8ab97ff0630465ee442fc58c1e4b1ab06f0301
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679163
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>