[build][rust] stop requiring output_name for rustc_binaries
This is Step 4 of the transition plan from https://fxbug.dev/64100

This makes it so that output_name defaults to the target name without
any `-` -> `_` remapping.

By now, we've been forcing output_name to be set to disambiguate between
the old behavior and the new behavior for a month, which is longer than
the week I wanted to wait (oops). So this change should be a no-op for
everyone who has updated their pending CLs sometime in the last month :)

Bug: 64100
Change-Id: I83c10bae9e797797f39b376be1881c5e36b70942
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680140
Commit-Queue: Dan Johnson <computerdruid@google.com>
Reviewed-by: Joseph Ryan <josephry@google.com>