[virtualization] Add Rust vsock device skeleton
This is the rough outline of the Rust vsock device (which is getting
large enough to warrant a review).

This is nearly entirely non-working, but is able to serve the relevant
FIDL protocols and accept connections.

Note that we're planning on deprecating the GuestVsockEndpoint protocol
once everything migrates to CFv2, and we'll serve GuestVsockAcceptor
directly and connect to HostVsockConnector.

Bug: 97355

Tested:
fx test virtio_vsock_unittests

Change-Id: Icc88b83b679d2fecaf2e593c3e42ec0eba8809b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668189
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: David Hastings <dahastin@google.com>