[fxfs] Roll more SuperBlock functionality into super_block.rs
This moves JournalMetrics (which only hold superblock metrics) into
super_block.rs as well as logic to load the freshest valid super block
and to save alternate A/B superblocks.

I have also added ObjectStore::new_root_parent() and attach_filesystem()
methods as part of this CL which deserves consideration.

I'm not a huge fan of these, but after trying a few things, this seemed
to be a straight-forward way avoid using an incomplete FxFilesystem with
a default SuperBlock to then read the actual SuperBlock and replace it.
The benefit of this should be greater in a subsequent change when SuperBlock
is moved from Journal to FxFilesystem.

I stopped short of completely internalizing SuperBlockInstance because
handles have to outlive the transaction they are used in which presents an
issue in init_empty(). They are also used in fsck.

Change-Id: I6507db4d3894f99b538d142ed8e4d626065999be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677866
Commit-Queue: Aaron Drew <ripper@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>