[compat] Only remove device when requested
Only remove the compat device if the driver owning the
device requested it to be removed.

The NodeController could have been unbound because
the child driver crashed (or was stopped). But in these
cases the DFv1 driver doesn't expect it's device to be
removed.

The DFv1 behavior would be to try and rebind the
child device up to 3 times, but for now we will do nothing.

Bug: 100470
Bug: 99885
Fixed: 98819

Change-Id: I405730639110396563046dcb36d6f5d749e8f0e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679005
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>