[assembly] Add create-flash-manifest operation
This operation currently reads the partitions config and exits, but will
soon contain the logic for constructing the flash manifest. I am running
the operation as part of the build to ensure that it continues working,
and will eventually replace the GN code for generating a flash manifest.

Bug: 97962
Change-Id: I9be3229ff1d6c309102007399dd8abe5e9442d03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678228
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Aidan Wolter <awolter@google.com>