[e2e] Adds some more error info into e2e tests
Workstation tests seem to be on the flakier side these days, but
the failure modes are somewhat opaque.  Adding these log messages
helped me pinpoint a number of interesting failure modes, as can
be seen in the attaached bug.

With this, I'd like to make the changes permanent, in the attempt
to reveal more of the failure modes.

Bug: 100441
Change-Id: Ifa28bcfede5a62fbe995a127225d567523a296c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/680204
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>