[testrunner] Use a temp out dir for host tests.
Sometimes a host test calling ffx doesn't properly shut the daemon down
and it might still be writing to the daemon log when we try to upload
the outputs to CAS with the swarming task outputs. This causes the bot
to fail as described in the attached bug. In this case, we should use a
temp out dir when running the test and move the contents to the actual
designated out dir once the test completes.

Bug: 99799
Change-Id: Ib0c80cb5dd23c77106a22f48fa18cfe9e1b28ca4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680306
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>