[kernel][scheduler] Fix Scheduler::RescheduleMask
Fix an issue in Scheduler::Reschedule mask which cause a local
preemption to fail to be recorded as pending when a thread's
preemption state's PreemptIsEnabled was false.  In some situations,
this could delay preemption on the local CPU incorrectly, in
particular, when a kernel thread dropped a contested Mutex.

Fixed: 100545
Change-Id: Ia7ae7c16bbddc08711b03cd24f9f64e6173dfa98
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680197
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: John Grossman <johngro@google.com>