[workshop] Add buggy persistence logic
Persists information in a slightly buggy way. Each profile key is saved
as json in a different file. These bugs will be used to showcase
diagnostics and testing.

There are a few bugs, but the deliberately added one works as follows:
1. A client opens a channel to profilestore.
2. The client creates a profile and makes modifications to it.
3. The client deletes the profile while keeping the channel created in
step 2 open.
4. The client closes the channel to the profile.

After this point, the handler in the server recreates the file,
recreating the deleted file.

Change-Id: I3ad99a21c30aa7dc1d276ea4e125eb1f30360de6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679507
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>