[credential_manager] Add CredentialReset Protocol.
This changeset introduces a new Reset method that simply calls
ResetTree and deletes all of the on-disk state. This is part of
a chain of CLs to land PinWeaver state being cleared on FDR.
Since Reset is only used in FDR and FDR doesn't need access to
mutate credentials it made sense to split this out into its
own distinct protocol that could be routed to FDR instead of
all the sensitive credential protocols.

Bug: 87687
Change-Id: I2918b2b24b79bf107850ab0045fb82597b3c64b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678647
API-Review: Jody Sankey <jsankey@google.com>
Commit-Queue: Ben Wright <benwright@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>