[fxt] Added Instant and Counter Events
This adds support for Instant and Counter Events. This also implements
serialization for Thread/String refs and Arguments as they are required
for events.

This events are implemented as described by the fxt documentation [1]

This is a reimplementation of the apis
ulib/trace-engine/context_api.cc[2] that are buffer and writer
independent.

[1] https://fuchsia.dev/fuchsia-src/reference/tracing/trace-format#event-record
[2] https://cs.opensource.google/fuchsia/fuchsia/+/master:zircon/system/ulib/trace-engine/context_api.cc;l=848;drc=0b95ec1c99b1378175e59f5a4f7ef79f6d484b0a

Change-Id: If41cad77bc363e9f62542f618303032cd78af9d4
Bug: 98177
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675286
Reviewed-by: Brian Hamrick <bhamrick@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>