[audio][tools] Fix 'virtual_audio --add'
Get virtual_audio_util working again. After recent changes,
these fixes are needed to reenable '--add', which is needed
for all virtual_audio scenarios.

The utility reveals that input device ring-buffers are kept
alive but stopped when the device is unplugged, but this is
NOT the case with outputs (an unplug entirely removes them).
In the test below, the '--write-rb' flag populates an input
ring-buffer, and successive concurrent wav_recorder-then-
wav_player invocations show the input ring-buffer to be
unavailable when initially added as unplugged, but then
available once plugged, and finally unavailable when
unplugged.

Test: virtual_audio --in --clear-format --add-format=6 \
                    --plug-props=2 --add    --wait     \
                    --plug --write-rb       --wait     \
                    --unplug                --wait

With the following before continuing each '--wait':
      wav_recorder --duration=3 --chans=1 --rate=8000  \
                   --int16 /tmp/wav.wav                \
      && wav_player /tmp/wav.wav

Change-Id: I90c7bd750349ccd78dd7cb54bc4da72798a37c3d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679722
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>