[wlan] Remove unused phyinfo from fullmac.
fuchsia.hardware.wlan.phyinfo only contains the old driver features
enum, which is about to be removed. The files in this change include
phyinfo but do not use anything from phyinfo. phyinfo will be
removed in a future change after all references to it are removed.

Bug: 88315
Test: fx test brcmfmac_test brcmfmac_sim_test wlanif_unittest
Change-Id: I1715770682f8224738a08c14989575b60bd7d8b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680863
Reviewed-by: Bjoern Johansson <bjoernj@google.com>
Commit-Queue: Karl Ward <karlward@google.com>