[media/vnext] add ffmpeg library
This CL adds a small library that makes ffmpeg easier to use for e.g.
decoders and demuxes. The code relating to AVCodecContext helps set
up such contexts for decoders. It also allows format information to
be easily extracted in the desired form from the codec contexts and
stream objects produced by the demux. The code relating to
AVIOContext marries the front end of the demux to the reader
abstraction defined in the reader library. Among other problems, the
issue of handling blocking read/seek calls using the asynchronous
reader abstraction is addressed.

Test: new unit tests
Change-Id: If110e77c5c952f2996bf122afa6e511a82ae8bbb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679996
Reviewed-by: Martin Puryear <mpuryear@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>