[FIDL] Create ServiceReconnector/ServiceHubConnector
ServiceReconnector can be used to easily stay connected to a FIDL
service and operate on it with the assumption that the service is always
connected.

ServiceHubConnector can be used to easily support the service hub fidl
pattern. It handles reconnecting to the service hub and the service, as
well as retrying failed calls upon request.

Change-Id: Ib26ef08b5afebdcd412ec56a084aaf74336fc438
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/653021
Commit-Queue: Zach Bush <zmbush@google.com>
Fuchsia-Auto-Submit: Zach Bush <zmbush@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>