[appmgr] Reject v1 manifests with non-canonical paths in sandbox
The sandbox section of CFv1 component manifests contains many path
elements specifying capabilities and subdirectories within capabilities
to map into the component's namespace. Internally appmgr handles many of
these paths with string manipulations and then passes them into fdio
routines to construct the namespace. If provided a non-canonical path,
when combined with a base component the constructed paths could
canonicalize to a surprising meeting. For example a "pkgfs" entry with
the non-canonical path "../boot" would turn into "/pkgfs/../boot" which
canonicalize to "/boot".  Currently appmgr relies on fdio_open_fd() not
canonicalizing the constructed paths in order to avoid escalation which
is fragile as other very similar-looking entry points do canonicalize
paths.

To avoid this issue, this change updates the component manifest parsing
logic to require that all entries within a "sandbox" section be valid
canonical Fuchsia paths. The only non-canonical path in the wild today
is a test case constructed to ensure that appmgr does not expand
capabilities when given such a path. This changes updates the test to
verify that a component with a non-canonical path does not launch.

Bug: 98543
Change-Id: I352f5cebfa750944930e122c2a12ab0c77d623fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678189
Reviewed-by: Ben Wright <benwright@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: James Robinson <jamesr@google.com>