[clippy] Fix lint deduplication
Previously we used the full rendered message to deduplicate lints,
but we found that the note indicating which category implied a lint
only gets added to the FIRST one emitted per clippy invocation.

Example:
```
= note: `--force-warn bare-trait-objects` implied by `--force-warn rust-2021-compatibility`
```

Since we do multiple invocations for toolchains and test variants,
the lint that comes first can change, and will have a different
rendered message content than the same lint in a different invocation,
causing both to show up in `fx clippy` output.

To fix this we're changing the key for deduplication from the rendered
text to a combination of the main message of a lint, it's "code" (e.g.
clippy::bad_bit_mask), and all of its primary spans, which should be
a more consistent indicator of uniqueness.

Change-Id: Ia82f696562d67323623e67574099714674b5020e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681347
Fuchsia-Auto-Submit: Joseph Ryan <josephry@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>