[wlan] Make sure NetworkPort waits for PortRemoved
The NetworkPort driver component would immediately destruct after
calling RemovePort. However the contract for netdevice requires that
resources associated with the port are not released until the removal
has been completed by a call to Removed. This lead to some rare crashes
when starting and stopping AP interfaces as netdevice would attempt to
make calls to the NetworkPort component after it was destroyed.

Add a completion and wait for it after calling RemovePort so that
Removed can finish and then signal the completion.

Bug: 100725
Test: Included unit test, manually run iperf3 while starting/stopping AP
Change-Id: I226f2568dffbaa82db09a877d33c76737743c6f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681348
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Bjoern Johansson <bjoernj@google.com>