[fidl] Remove usages of coding tables in LLCPP
This CL removes coding tables from the LLCPP type traits and usages of
the coding tables.

However, there are some usages of coding tables that are currently
generated on LLCPP types but are not used in LLCPP. A new trait was
created so the coding tables can still be used in the near term, but do
not increase the size of regular LLCPP usages.

Binary sizes decreased by 300-450k!

Change-Id: I4657ba2342783b2c1969f44b69509827d2e47839
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680463
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>