[wlan] Remove unused phyinfo from wlan/lib/common.
fuchsia.hardware.wlan.phyinfo only contains the old driver features
enum, which is about to be removed. The files in this change include
phyinfo but do not use anything from phyinfo. phyinfo will be
removed in a future change after all references to it are removed.

Bug: 88315
Test: fx test brcmfmac_test brcmfmac_sim_test wlan-hw-sim-test \
              wlan-mlme-tests wlan_tests wlanstack-tests
Change-Id: I57aeac52b3f280e112af7582000782779dfc7070
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680150
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Commit-Queue: Karl Ward <karlward@google.com>