[scrutiny][update-package] Inspect images JSON file
This change modifies how the scrutiny core packages collector
inspects the update package. Previously, the collector looked for
an entry named "zbi.signed" or "zbi" in the update package. Now,
the collector looks for said entry and checks for the same entry
in the "fuchsia" slot of the new images JSON file that is expected
to be included in the update package.

Bug: 98030
Change-Id: I50ae94485db4f56be301200de9ba3347d8c2d994
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678358
Commit-Queue: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Ben Wright <benwright@google.com>