[iwlwifi] Fix WPA2 on NUC
After the firmware decrypts the packet, it leaves a useless
CCMP header. The driver has to remove it before it is sent
to the MLME.

This CL re-uses the iwl_mvm_create_packet() function in mvm/rxmq.c.

Tested: On NUC and Atlas, tested on open network and WPA2 APs.
        Also added a test case to cover iwl_mvm_create_packet().
Bug: 88353

Change-Id: I41a2e4787de38997cfd54c72a9e9d5ec6b355d99
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680604
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Yung-chieh Lo <yjlou@google.com>