[wlan] Remove priority_queue internal pop vector
The priority_queue driver component had an internal vector used for
popping and it would then return a span over that vector. This required
a caller to prevent other threads from popping while the span was being
used. This extended locking requirement is very inefficient because it
significantly extends the period during which other threads can push
onto the queue.

Remove the internal pop vector and require that the caller provide a
FrameContainer that the popped frames are placed into.

Test: Included unit tests, iperf3 while starting/stopping AP interfaces.
Bug: 100734
Change-Id: I6c2e07145debe7f830c95d0023d5ed950c839381
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681349
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>