[netstack3] Include addresses in socketmap tags
This addresses a shortcoming in the original socketmap design. The
assumption there was that all addresses that would conflict would also
have a shadowing relationship in one direction or another. This is not
the case, in part due to the inclusion of device IDs in address vectors.

A UDP socket listening on 0.0.0.0:8080 and bound to device D, for
example, is expected to prevent the binding of another listening socket
on 127.0.0.1:8080 (no device). The address vectors for both these
sockets shadow 0.0.0.0:8080 (no device) but there is no direct shadowing
relationship.

Change-Id: Ia094e63ba3448bd27ee72c9ea02fdb2011f039b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680114
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>