[ffx] Treat "no ssh port" as port 22 for filtering
Changes the logic of the "matching port" subclause of the target
collection matching code to treat 22 as a default for a target when
matching against port 22 is requested.

There's potential here for a larger, deeper, fix to how mdns targets
are added to the collection, since they probably *should* have port 22
specified as their port, but I think it's best to save that for another
day. For now, this enables using `ffx target` commands with a target that
includes the port number when those hosts were discovered rather than
explicitly added and had an ssh port of 22.

Also updated the target fidl documentation to reflect the existing
reality along with these updates. It seems to have been written when
this protocol was much more restrictive.

Fixed: fxbug.dev/98589
Test: fx test host_x64/ffx_daemon_target_lib_test

Change-Id: I2a5cf1165db681887f6c93bb2518d51dbb22db72
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681344
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
API-Review: Jeremy Manson <jeremymanson@google.com>