[fdf][compat] Try and find a real sysmem fragment...
Before returning the fake one.

Previously we would unconditionally return the fake sysmem fragment.
This CL changes the logic so that we always try the "real" GetProtocol
first, but if that fails and the caller was looking for sysmem, we
return our fake sysmem fragment.

Test: display works on atlas on dfv2
Test: display works on optiplex on dfv2 (with followup simple-intel cl)
Bug: 100807
Change-Id: Ic06679933092559c43b6a153536b3b40a57583b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681406
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>