[fdf] Split out driver_host.fidl from fdf FIDL
The fdf FIDL is part of the SDK. The driver host FIDL is only used
internally and therefore not needed in the SDK. This change splits it
out into its own FIDL library and refactors all usages of it. The
DriverStartArgs gets separated out from driver_host.fidl and left in
fdf since it is used in the SDK.

Bug: 97984
Tests: Existing unit and integration tests work. New tests are
not needed as there are no changes in behavior.

Change-Id: I4ce005c32032c82fd9cfd42d449a74c9b9ff80bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681663
Reviewed-by: Sarah Chan <spqchan@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>