[input-synthesis] Test handling newlines and tabs
I noticed that we never tested input synthesis for nonprintable
chars like \n and \t.  I added this to check that we're doing
the right thing, and it seems that we are.

Added pretty assertions too for easier checking of mismatches.

Then it occurred to me that it would be just fine to merge this.

Bug: 79807

Change-Id: I995fc71f4f6edf0b90e0d1e50f1937d3516c8006
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680408
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>