[brcmfmac] Synchronously remove netdev port
The network device port was being removed as part of the destruction.
The destruction happens as a result of async remove on the wlanif
device. Unfortunately it seems that new interfaces can be complete
before that destruction has completed successfully. The interface was
destroyed in firmware however so firmware would reuse the interface ID.
When attempting to use that interface ID as the port ID for the
netdevice port the port creation would fail because a port with that ID
already existed, it hadn't been remove completely yet.

Fix this by synchronously removing the port in the call to destroy the
interface, before async remove is called.

Bug: 100854
Test: Manual testing, verify driver handles interfaces up and down
Change-Id: Iaf53386dd63ffe4df1a2d226f5cf139fecd914c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681906
Fuchsia-Auto-Submit: Bjoern Johansson <bjoernj@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>